Added Panel, sub102 and mixer now use this for UI.
Changes to darkbat api sender to allow bigger shapes.

Added more controls to sub102.



git-svn-id: file:///home/pi/svn/RS7@47 04a3551b-0df6-43f4-9fb6-bc03da9130d0

+++ ALSAWriterImpl.cpp
@@ -67,7 +67,7 @@ printf("GDR: opened sound\n");
   snd_pcm_hw_params_get_period_size(params, &frames,
                                     &dir);
   size = frames * 4; /* 2 bytes/sample, 2 channels */
- fprintf(stderr, "GDR: buffer size in periods=%d\n", frames );
+ fprintf(stderr, "GDR: buffer size in periods=%d\n", (int)frames );
   return frames;
 }
